requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
PySocks==1.7.1
pytest==8.0.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
